requirements.txt
uvicorn~=0.30.6
DrissionPage~=4.0.5.6
fastapi~=0.112.1
beautifulsoup4~=4.12.3
pyyaml~=6.0
fastapi_cache2~=0.2.2
fastapi-utils~=0.7.0
typing-inspect~= 0.9.0
